Highlighting and color palette updates
