+++ NineMensMorris/Style.java
@@ -1,0 +1,10 @@
+package NineMensMorris;
+
+import javafx.scene.paint.Color;
+
+public class Style {
+	public static Color lightBlue = Color.web("#afc1cc");
+	public static Color midBlue = Color.web("#7895a2");
+	public static Color darkBlue = Color.web("#517281");
+	public static Color darkRed = Color.web("#ff3c38");
+}
